+++ src/components/index.js
@@ -2,3 +2,5 @@ export { default as CustomSwitch } from './CustomSwitch';
 export { default as CustomInput } from './CustomInput';
 export { default as Spinner } from './Spinner';
 export { default as Navigation } from './Navigation';
+export { default as GeneralError } from './GeneralError';
+export { default as GeneralAuth } from './GeneralAuth';