+++ src/views/ValidateCredentials/components/Register.js
@@ -1,30 +1,28 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
 
-import { Button, Typography, makeStyles, Grid } from '@material-ui/core';
+import { Button, makeStyles, Hidden } from '@material-ui/core';
 import ArrowBackIosRoundedIcon from '@material-ui/icons/ArrowBackIosRounded';
 import LockRoundedIcon from '@material-ui/icons/LockRounded';
 
+import { GeneralAuth } from 'components';
+import registerAuthenticationSvg from 'assets/images/undraw/register_authentication.svg';
+
 const useStyles = makeStyles((theme) => ({
-  mainContent: {
-    padding: '170px 150px',
-  },
-  mainTitle: {
-    fontSize: '4em',
-  },
-  secondaryTitle: {
-    fontSize: '2em',
-    margin: '1rem 0',
-  },
-  subtitle: {
-    fontSize: '1.2em',
-  },
   buttonsSection: {
-    marginTop: '1rem',
+    marginTop: 20,
     display: 'flex',
     justifyContent: 'space-between',
+    [theme.breakpoints.down('md')]: {
+      justifyContent: 'space-between',
+      flexDirection: 'column',
+    },
   },
-  returnHomeButton: {},
+  returnHomeButton: {
+    [theme.breakpoints.down('md')]: {
+      marginTop: 20,
+    },
+  },
   registerButton: {
     color: theme.palette.white,
   },
@@ -39,23 +37,19 @@ const Register = () => {
   };
 
   return (
-    <Grid container className={classes.mainContent}>
-      <Grid item xs></Grid>
-      <Grid item xs>
-        <div>
-          <Typography variant="h1" className={classes.mainTitle}>
-            ¡Enhorabuena!
-          </Typography>
-          <Typography variant="h2" className={classes.secondaryTitle}>
-            Te encuentras inscrito en el padrón electoral
-          </Typography>
-          <Typography variant="subtitle1" className={classes.subtitle}>
-            Para generar tu contraseña, haz click en el siguiente botón.
-          </Typography>
-          <Typography variant="subtitle1" className={classes.subtitle}>
-            Te enviaremos la contraseña a tu correo institucional.
-          </Typography>
-        </div>
+    <GeneralAuth
+      mainTitle="
+¡Enhorabuena!"
+      secondaryTitles={[
+        'Usted se encuentra inscrito en el padrón electoral.',
+        'Para poder generar su contraseña, haga click en "Generar contraseña".',
+      ]}
+      subtitle="Le enviaremos su contraseña a su correo institucional.
+"
+      altImage="register_authentication_svg"
+      srcImage={registerAuthenticationSvg}
+    >
+      <Hidden mdDown>
         <div className={classes.buttonsSection}>
           <Button
             className={classes.returnHomeButton}
@@ -78,8 +72,33 @@ const Register = () => {
             Generar contraseña
           </Button>
         </div>
-      </Grid>
-    </Grid>
+      </Hidden>
+
+      <Hidden lgUp>
+        <div className={classes.buttonsSection}>
+          <Button
+            className={classes.registerButton}
+            endIcon={<LockRoundedIcon />}
+            size="large"
+            variant="contained"
+            color="primary"
+            onClick={returnToHome}
+          >
+            Generar contraseña
+          </Button>
+          <Button
+            className={classes.returnHomeButton}
+            startIcon={<ArrowBackIosRoundedIcon />}
+            size="large"
+            variant="outlined"
+            color="primary"
+            onClick={returnToHome}
+          >
+            Regresar al inicio
+          </Button>
+        </div>
+      </Hidden>
+    </GeneralAuth>
   );
 };
 export default Register;