+++ src/views/ValidateCredentials/components/Unauthorized.js
@@ -1,77 +1,71 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
 
-import { Button, Typography, makeStyles, Grid } from '@material-ui/core';
+import {
+  Button,
+  makeStyles,
+  Typography,
+  useMediaQuery,
+  useTheme,
+} from '@material-ui/core';
 import ArrowBackIosRoundedIcon from '@material-ui/icons/ArrowBackIosRounded';
 
+import { GeneralAuth } from 'components';
+import notAuthorizedInRegister from 'assets/images/undraw/not_authorized_in_register.svg';
+
 const useStyles = makeStyles((theme) => ({
-  mainContent: {
-    padding: '170px 150px',
-  },
-  mainTitle: {
-    fontSize: '4em',
-  },
-  secondaryTitle: {
-    fontSize: '2em',
-    margin: '1rem 0',
-  },
-  subtitle: {
-    fontSize: '1.2em',
-  },
   buttonsSection: {
     marginTop: '1rem',
-  },
-  returnHomeButton: {},
-  registerButton: {
-    color: theme.palette.white,
   },
   emailTo: {
     color: theme.palette.black,
     fontWeight: 'bold',
+    fontSize: '1.2em',
+    marginTop: '1rem',
   },
 }));
 
 const Unauthorized = () => {
   const classes = useStyles();
   const history = useHistory();
+  const theme = useTheme();
+  const isMobile = useMediaQuery(theme.breakpoints.down('xs'));
 
   const returnToHome = () => {
     history.push('/');
   };
 
   return (
-    <Grid container className={classes.mainContent}>
-      <Grid item xs></Grid>
-      <Grid item xs>
-        <div>
-          <Typography variant="h1" className={classes.mainTitle}>
-            ¡Oh, vaya!
-          </Typography>
-          <Typography variant="h2" className={classes.secondaryTitle}>
-            Al parecer no te encuentras inscrito en el padrón electoral
-          </Typography>
-          <Typography variant="subtitle1" className={classes.subtitle}>
-            Si crees que es un error, por favor contactanos a{' '}
-            <a className={classes.emailTo} href="mailto:ceuni@uni.edu.pe">
-              ceuni@uni.edu.pe
-            </a>
-            .
-          </Typography>
-        </div>
-        <div className={classes.buttonsSection}>
-          <Button
-            className={classes.returnHomeButton}
-            startIcon={<ArrowBackIosRoundedIcon />}
-            size="large"
-            variant="outlined"
-            color="primary"
-            onClick={returnToHome}
-          >
-            Regresar al inicio
-          </Button>
-        </div>
-      </Grid>
-    </Grid>
+    <GeneralAuth
+      mainTitle="
+      ¡Oh, vaya!"
+      secondaryTitles={[
+        'Usted no se encuentra inscrito en el padrón electoral',
+      ]}
+      subtitle="Si crees que es un error, por favor no dude contactarnos a:"
+      altImage="not_authorized_in_register_svg"
+      srcImage={notAuthorizedInRegister}
+    >
+      <Typography variant="subtitle1">
+        <a className={classes.emailTo} href="mailto:ceuni@uni.edu.pe">
+          ceuni@uni.edu.pe
+        </a>
+      </Typography>
+
+      <div className={classes.buttonsSection}>
+        <Button
+          fullWidth={isMobile}
+          className={classes.returnHomeButton}
+          startIcon={<ArrowBackIosRoundedIcon />}
+          size="large"
+          variant="outlined"
+          color="primary"
+          onClick={returnToHome}
+        >
+          Regresar al inicio
+        </Button>
+      </div>
+    </GeneralAuth>
   );
 };
 export default Unauthorized;